botlib: fix FuzzyWeight family return type double -> float (was IDA mistype)
FuzzyWeight_r, FuzzyWeightUndecided_r, FuzzyWeight, and
FuzzyWeightUndecided were all declared returning double, but disasm
evidence (raw objdump opcode bytes, PE and ELF originals both) shows
(single-precision, 4-byte, 0xd9 prefix) form, never QWORD (0xdd,
double). Q3A's be_ai_weight.c confirms the same four functions as
float in the same-author reference source. This was a pure IDA
return-type mis-inference carried through the whole reconstruction
(a recursive ST(0)-returning function without checking operand
widths), not an authentic 1999 typing choice.

Fixing the prototype resolved a caller-side fstp+fld double-to-float
rounding roundtrip that was corrupting several callers' codegen:

  FuzzyWeight_r:            44 insn_diffs -> FULL MATCH
  FuzzyWeightUndecided_r:  179 insn_diffs -> FULL MATCH
  FuzzyWeight:              92 insn_diffs -> FULL MATCH
  BotChooseLTGItem:        256 -> 254 insn_diffs (ripple)
  BotChooseNBGItem:        239 -> 237 insn_diffs (ripple)

Also fixes BotChooseBestFightWeapon (46 -> 4 insn_diffs): once the
FuzzyWeight prototype was corrected, reordered the three independent
before the wc load) and moved the bestweight declaration to match
gcc's actual stack-slot assignment and instruction scheduling for
this statement set, both confirmed against the real disassembly.
Residual 4 insn_diffs is a numweapons>0 comparison where real
materializes zero via a coincidental memory-reuse trick instead of
an immediate compare -- register-allocator internals, not
source-expressible.

PE-oracle: all seven affected functions confirmed FULL MATCH; full
810-function audit_dll.sh ripple check is byte-for-byte identical to
the committed baseline. The double->float fix is a pure ELF win with
zero PE impact.

Source-commit: b517c49025b53a84e66b1198843c71f9df7bd178

+++ botlib/be_ai_weight.c
@@ -541,7 +541,7 @@ int __cdecl FindFuzzyWeight(weightconfig_t *wc, const char *name)
 // Recursive tree walk for the discrete case: tests `facts[index]` against
 // each seperator's threshold, descends the matching branch, returns the
 // leaf weight. LIVE: core of FuzzyWeight.
-double __cdecl FuzzyWeight_r(int *inventory, fuzzyseperator_t *fs)
+float __cdecl FuzzyWeight_r(int *inventory, fuzzyseperator_t *fs)
 {
   float scale, w1, w2;
 
@@ -575,7 +575,7 @@ double __cdecl FuzzyWeight_r(int *inventory, fuzzyseperator_t *fs)
 // Recursive tree walk with fuzzy interpolation: when a fact straddles a
 // seperator threshold, blends the two branch weights instead of snapping.
 // LIVE: core of FuzzyWeightUndecided (used for smooth priority changes).
-double __cdecl FuzzyWeightUndecided_r(int *inventory, fuzzyseperator_t *fs)
+float __cdecl FuzzyWeightUndecided_r(int *inventory, fuzzyseperator_t *fs)
 {
   float scale, w1, w2;
 
@@ -609,18 +609,18 @@ double __cdecl FuzzyWeightUndecided_r(int *inventory, fuzzyseperator_t *fs)
 // Public entry: discrete fuzzy weight for one weight_t (one item/weapon).
 // Returns the leaf value of the decision tree given current bot facts.
 // LIVE: called by the item/weapon goal scorers every think.
-double __cdecl FuzzyWeight(int *facts, weight_t *w)
+float __cdecl FuzzyWeight(int *facts, weight_t *w)
 {
-  /* Thin wrapper over FuzzyWeight_r, which returns a double in ST(0). */
+  /* Thin wrapper over FuzzyWeight_r, which returns a float in ST(0). */
   return FuzzyWeight_r(facts, w->firstseperator);
 }
 //----- (10036CA0) --------------------------------------------------------
 // Public entry: fuzzy-interpolated weight for one weight_t. Smooth
 // variant of FuzzyWeight; avoids snap behaviour when a fact value
 // crosses a threshold. LIVE.
-double __cdecl FuzzyWeightUndecided(int *facts, weight_t *w)
+float __cdecl FuzzyWeightUndecided(int *facts, weight_t *w)
 {
-  /* Binary at 0x10036CA0 is a thin wrapper around FuzzyWeightUndecided_r (returns double). */
+  /* Binary at 0x10036CA0 is a thin wrapper around FuzzyWeightUndecided_r (returns float). */
   return FuzzyWeightUndecided_r(facts, w->firstseperator);
 }
 //----- (10036CD0) --------------------------------------------------------
